+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - added truncateToolResult to extension context to allow extensions to use it to truncate the tool results
 - added truncation of non-power tool results on smart compaction
 - replaced sharp native dependency with empty stub to prevent SIGILL on older CPUs
+- added truncation of read file tool results on smart compaction
 
 ## [0.66.0]